third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/serviceworker/fetch-cors-xhr-expected.txt

CONSOLE WARNING: line 14: Subresource requests whose URLs contain embedded credentials (e.g. `https://user:pass@host/`) are deprecated, and will be blocked in M59, around June 2017. See https://www.chromestatus.com/feature/5669008342777856 for more details.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?ignore. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?ignore. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?ACAOrigin=http://127.0.0.1:8000&ignore. The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the request's credentials mode is 'include'. Origin 'http://127.0.0.1:8000' is therefore not allowed access. The credentials mode of requests initiated by the XMLHttpRequest is controlled by the withCredentials attribute.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?Auth&ignore. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access. The response had HTTP status code 401.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?Auth&ignore. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?Auth&ACAOrigin=http://127.0.0.1:8000&ignore. The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the request's credentials mode is 'include'. Origin 'http://127.0.0.1:8000' is therefore not allowed access. The credentials mode of requests initiated by the XMLHttpRequest is controlled by the withCredentials attribute.
This is a testharness.js-based test.
PASS Verify CORS XHR of fetch() in a Service Worker 
Harness: the test ran to completion.
